backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
redis
requests
google-generativeai
sentence-transformers
qdrant-client
loguru
pydantic
